bucketlist/requirements.txt
appnope==0.1.0
coverage==4.0.3
coveralls==1.1
decorator==4.0.6
Django==1.9.1
django-rest-swagger==0.3.4
djangorestframework==3.3.2
docopt==0.6.2
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
PyYAML==3.11
requests==2.9.1
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
